refactor: rename ShareableIdentifiers class
